Fix: updated authorsation

+++ controllers/album-controller.js
@@ -69,6 +69,28 @@ exports.create = async (req, res) => {
 exports.update = async (req, res) => {
   const requestBody = await validateUpdate(req.body);
 
+  const { user } = req;
+  // is Owner : check if it is the artist
+  const isArtist = await Artist.findOne({
+    user: user._id, _id: req.params.artistId
+  });
+  if (!isArtist) {
+    return res.status(401).send({
+      success: false, 
+      message: 'Permission denied or artist does not exist'
+    });
+  }
+
+  const owner = await Album.findOne({
+    artist: req.params.artistId, _id: req.params.id
+  });
+  if (!owner) {
+    return res.status(401).send({
+      success: false,
+      message: 'Permission denied or artist does not exist'
+    });
+  }
+
   // we want to make upload
   if (('file' in req)) {
     // in the future, try to
@@ -98,6 +120,30 @@ exports.update = async (req, res) => {
  * Delete an album
  */
 exports.delete = async (req, res) => {
+
+  
+  const { user } = req;
+  // is Owner : check if it is the artist
+  const isArtist = await Artist.findOne({
+    user: user._id, _id: req.params.artistId
+  });
+  if (!isArtist) {
+    return res.status(401).send({
+      success: false, 
+      message: 'Permission denied or artist does not exist'
+    });
+  }
+
+  const owner = await Album.findOne({
+    artist: req.params.artistId, _id: req.params.id
+  });
+  if (!owner) {
+    return res.status(401).send({
+      success: false,
+      message: 'Permission denied or album does not exist'
+    });
+  }
+
   const filter = { _id: req.params.id, artist: req.params.artistId };
 
   const album = await Album.findOneAndRemove(filter);
